% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/0functions.R
\name{make.tapersetup}
\alias{make.tapersetup}
\title{Create a taper setup for a SPAM matrix}
\usage{
make.tapersetup(d, f.tap, taprange)
}
\arguments{
\item{d}{A \code{matrix} that is a symmetric distance matrix.}

\item{f.tap}{A \code{function} that is the tapering function to use.}

\item{taprange}{A \code{double} that is used to specify the gamma.}
}
\value{
A \code{list}
\itemize{
\item \code{n} - row/col of matrix
\item \code{good.dists} - distances that meet tapering requirements
\item \code{taps} - results of applying the taper function
\item \code{ja} - ordered column indices
\item \code{ia} -  pointer to the beginning of each row in the arrays entries and colindices
\item \code{index} - element position in matrix
}
}
\description{
Get the storage information we'll need to create a sparse matrix
}
\details{
This function is meant to be used to create SPAM matrices.
Do not use output for Eigen matrices.
}
\examples{
data(anom1962)
d = rdist.earth(head(loc))
make.tapersetup(d, wendland2.1, 20)
}

