% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/0functions.R
\name{nll.1taper}
\alias{nll.1taper}
\title{Second Tapering Implementation}
\usage{
nll.1taper(x, setup, z)
}
\arguments{
\item{x}{A \code{double} indicating the rho, or scaling parameter}

\item{setup}{Results returned from make.tapersetup}

\item{z}{A \code{vec} that contains the finite number of observations}
}
\description{
The function implements the second tapering method described
in the Covariance Tapering paper.
}
\examples{
data(anom1962)
d = rdist_earth1(loc)
setup.eigen = make.tapersetup(d,taprange = 50)
val = nll.1taper(20, setup, z)
}

