% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/0functions.R
\name{nll.2taper}
\alias{nll.2taper}
\title{Second Tapering Implementation}
\usage{
nll.2taper(x, setup, z)
}
\arguments{
\item{x}{A \code{double} indicating the rho, or scaling parameter}

\item{setup}{Results returned from make.tapersetup}

\item{z}{A \code{vec} that contains the finite number of observations}

\item{rescol}{An \code{int} that contains how many non-zero elements should be in a column.}
}
\value{
A \code{double}
}
\description{
The function implements the second tapering method described
in the Covariance Tapering paper.
}
\examples{
data(anom1962)
d = rdist_earth1(loc)
setup.eigen = make_tapersetup_eigen(d,taprange = 50)
val = nll_2taper(20, setup.eigen$n,
}

