% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/0functions.R
\name{nll}
\alias{nll}
\title{Create a taper setup for an Eigen sparse matrix}
\usage{
nll(x)
}
\arguments{
\item{x}{A \code{double} that represents rho, which is the scaling parameter of the exponential covariance function}
}
\value{
A \code{double}
}
\description{
Generate the storage information we'll need to create a sparse eigen matrix
}
\examples{
data(anom1962)
d = rdist_earth1(head(loc))
nll_arma(20)
}

