% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nll_1taper}
\alias{nll_1taper}
\alias{nll_1taper_parallel}
\title{First Tapering Implementation}
\usage{
nll_1taper(x, n, good_dists, taps, ia, ja, z, rescol)

nll_1taper_parallel(x, n, good_dists, taps, ia, ja, z, rescol, cores)
}
\arguments{
\item{x}{A \code{double}}

\item{n}{A \code{double} that indicates the matrix row/col.}

\item{good_dists}{A \code{VectorXd} that contains the distances which meet tapering requirements}

\item{taps}{A \code{VectorXd} results of applying the taper function}

\item{ia}{A \code{VectorXd} containing the ordered row indices}

\item{ja}{A \code{VectorXd} containing the ordered column indices}

\item{z}{A \code{VectorXd} element position in matrix}

\item{rescol}{An \code{int} that contains how many non-zero elements should be in a column.}

\item{cores}{An \code{int} that indicates the number of cores to divide the task over.}
}
\value{
A \code{double}
}
\description{
The function implements the first tapering method described
in the Covariance Tapering paper.
}
\details{
Currently the EIGEN sparse matrix cholesky decomposition differs from that of SPAM.
As a result, the forward solve step finds a different vector. This in turn returns a higher distval.
}
\section{Functions}{
\itemize{
\item \code{nll_1taper_parallel}: 
}}
\examples{
data(anom1962)
d = rdist_earth1(loc)
setup.eigen = make_tapersetup_eigen(d,taprange = 50)
val = nll_1taper(20, setup.eigen$n,
                 setup.eigen$good.dists,
                 setup.eigen$taps,
                 setup.eigen$ia,
                 setup.eigen$ja,
                 z,
                 setup.eigen$rescol)
}

