% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nll_arma}
\alias{nll_arma}
\alias{nll_parallel}
\title{Create a taper setup for an Eigen sparse matrix}
\usage{
nll_arma(x, d, z, n)

nll_parallel(x, d, z, n, cores = 1L)
}
\arguments{
\item{x}{A \code{double} that represents rho, which is the scaling parameter of the exponential covariance function}

\item{d}{A \code{matrix} that is a symmetric distance matrix.}

\item{z}{A \code{vec} that contains the finite number of observations}

\item{n}{A \code{double} that indicates the number of rows/cols of the matrix.}

\item{cores}{A \code{integer} representing how many cores should be loaded}
}
\value{
A \code{double}
}
\description{
Generate the storage information we'll need to create a sparse eigen matrix
}
\section{Functions}{
\itemize{
\item \code{nll_parallel}: 
}}
\examples{
data(anom1962)
d = rdist_earth1(head(loc))
nll_arma(20, d, z, n)
}

