% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spatialtalk.R
\docType{package}
\name{spatialtalk-package}
\alias{spatialtalk-package}
\title{spatialtalk Package}
\description{
spatialtalk contains components from my talk
}
\details{
\tabular{ll}{
Package: \tab spatialtalk\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-05-12\cr
License: \tab MIT + file LICENSE\cr
}
Some details should be added
}
\examples{
# Fill in at a later time
}
\author{
James Balamuta \email{balamut2@illinois.edu}

James Balamuta \email{balamut2@illinois.edu}
}

