% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{wendland2_1}
\alias{wendland2_1}
\title{Taper Function}
\usage{
wendland2_1(d, taprange)
}
\arguments{
\item{d}{A \code{vec} containing the small distance values}

\item{taprange}{A \code{double} indicating the taper range}
}
\value{
A \code{vec} that has had tapering applied to d.
}
\description{
Taper function we're using
}
\examples{
data(anom1962)
d = rdist_earth1(head(loc))
wendland2_1(d, 20)
}

